Preserve default date range on chart.
Changed so that if the chart page is brought up with the default date range, it refreshrd itself using the default range rather than setting a particular ending date. (That was a bug.) This way the 'right end' of the chart is always current, even after it's been refreshed many times over many hours. If the user manually selects a specific date range, however, that range will be used for every refresh after that point.